+++ zview/plugins/seq/seq.c
@@ -699,7 +699,7 @@ boolean __CDECL reader_read(IMGINFO info, uint8_t *buffer)
 	x = 0;
 	if (cur_y == 0)
 	{
-		int handle = info->_priv_var_more;
+		int handle = (int)info->_priv_var_more;
 		Fseek(frame_positions[info->page_wanted], handle, SEEK_SET);
 		if ((size_t)Fread(handle, sizeof(frame_header), &frame_header) != sizeof(frame_header))
 			return FALSE;
